resume_screening_app/requirements.txt
streamlit
langchain
langchain-google-genai
langchain-community
langchain-chroma
chromadb
python-dotenv
pypdf
python-docx
pandas
